Update RustBuffer logic to use `uint64_t`

+++ bindgen/src/bindings/cpp/templates/callback_conv.hpp
@@ -6,7 +6,7 @@ struct {{ typ|ffi_converter_name }} {
     static uint64_t lower(const {{ type_name }} &);
     static {{ type_name }} read(RustStream &);
     static void write(RustStream &, const {{ type_name }} &);
-    static int32_t allocation_size(const {{ type_name }} &);
+    static uint64_t allocation_size(const {{ type_name }} &);
 
     inline static HandleMap<{{ canonical_type_name }}> handle_map = {};
 };
